stub: workaround race in sysfs attribute access
In stress testing detach/attach I noticed a race in
the access of the sysfs attributes which can make 
usbipd fail to export the device.

When a client detaches, the stub driver removes the
sysfs attributes (stub_disconnect()) and recreates 
them shortly after (stub_probe()).

During that time any attempts to attach the device
will fail since usbipd tries but cannot read the 
sysfs attribute.

This is a problem for quick successive detach and
attach. This was noticed in a real setup as well,
although infrequently.

The client can't tell that this is a temporary error
and that it would succeed it if just retried shortly
after - so it seems best to work around it in usbipd.

This patch makes usbip poll waiting for the sysfs to
appear if it does not exist. 

The number of retries is obviously somewhat arbitrary
and probably higher than necessary but should result 
in a delay that is still tolerable in the worst case.

The maximum number of required retries I saw on a 
heavily loaded system were 43.

So while obviously not clean and pretty solution, I
think this workaround is the best we can do with the
current sysfs interface.



git-svn-id: https://usbip.svn.sourceforge.net/svnroot/usbip@155 88f907b9-6f29-0410-9914-a5135e967dcd